Convert list from xml to data.frame without changes

Convert list from xml to data.frame without changes

Usage

cbr_security_info(
  name,
  firstDate = as.Date("2000-01-01"),
  secondDate = Sys.Date()
)

Arguments

name

the name of variable from https://cbr.ru/secinfo/secinfo.asmx

firstDate

start of period, Date, as.Date("2000-01-01") by default

secondDate

end of period, Date, Sys.Date() by default

Value

data.frame

Examples

df <- cbr_security_info("AuctionsXML")
#> StartTag: invalid element name
#> StartTag: invalid element name
#> Opening and ending tag mismatch: link line 25 and head
#> StartTag: invalid element name
#> StartTag: invalid element name
#> StartTag: invalid element name
#> StartTag: invalid element name
#> StartTag: invalid element name
#> StartTag: invalid element name
#> Opening and ending tag mismatch: img line 124 and div
#> StartTag: invalid element name
#> StartTag: invalid element name
#> StartTag: invalid element name
#> StartTag: invalid element name
#> StartTag: invalid element name
#> StartTag: invalid element name
#> Specification mandates value for attribute data-popup-cookies
#> StartTag: invalid element name
#> StartTag: invalid element name
#> StartTag: invalid element name
#> StartTag: invalid element name
#> StartTag: invalid element name
#> StartTag: invalid element name
#> StartTag: invalid element name
#> StartTag: invalid element name
#> StartTag: invalid element name
#> StartTag: invalid element name
#> StartTag: invalid element name
#> StartTag: invalid element name
#> Specification mandates value for attribute data-popup-cookies-btn
#> Opening and ending tag mismatch: div line 121 and body
#> Opening and ending tag mismatch: body line 92 and html
#> Premature end of data in tag link line 21
#> Error: 1: StartTag: invalid element name
#> 2: StartTag: invalid element name
#> 3: Opening and ending tag mismatch: link line 25 and head
#> 4: StartTag: invalid element name
#> 5: StartTag: invalid element name
#> 6: StartTag: invalid element name
#> 7: StartTag: invalid element name
#> 8: StartTag: invalid element name
#> 9: StartTag: invalid element name
#> 10: Opening and ending tag mismatch: img line 124 and div
#> 11: StartTag: invalid element name
#> 12: StartTag: invalid element name
#> 13: StartTag: invalid element name
#> 14: StartTag: invalid element name
#> 15: StartTag: invalid element name
#> 16: StartTag: invalid element name
#> 17: Specification mandates value for attribute data-popup-cookies
#> 18: StartTag: invalid element name
#> 19: StartTag: invalid element name
#> 20: StartTag: invalid element name
#> 21: StartTag: invalid element name
#> 22: StartTag: invalid element name
#> 23: StartTag: invalid element name
#> 24: StartTag: invalid element name
#> 25: StartTag: invalid element name
#> 26: StartTag: invalid element name
#> 27: StartTag: invalid element name
#> 28: StartTag: invalid element name
#> 29: StartTag: invalid element name
#> 30: Specification mandates value for attribute data-popup-cookies-btn
#> 31: Opening and ending tag mismatch: div line 121 and body
#> 32: Opening and ending tag mismatch: body line 92 and html
#> 33: Premature end of data in tag link line 21